新增起始点在岛洞内或外部的情况+结束点不在地图内的情况
新增起始点在岛洞内或外部的情况+结束点不在地图内的情况

+++ Base/NavMeshHelper.h
@@ -13,7 +13,7 @@ namespace NavMeshBase
         {
             double S = 0;
             
-            int n = points.size();
+            size_t n = points.size();
             for (int i = 0; i < n - 1; i++) {
                 const Vector3 point = points[i];
                 const Vector3 nextPoint = points[i + 1];
